+++ src/main/java/com/github/thesench/solr/dsl/stream/expr/decorators/OuterHashJoin.java
@@ -1,0 +1,25 @@
+package com.github.thesench.solr.dsl.stream.expr.decorators;
+
+import com.github.thesench.solr.dsl.stream.expr.params.Hashed;
+import com.github.thesench.solr.dsl.stream.expr.params.On;
+
+import org.apache.solr.client.solrj.io.stream.expr.StreamExpression;
+
+public class OuterHashJoin {
+    private OuterHashJoin() {}
+
+    /**
+     * @see <a href="https://solr.apache.org/guide/8_10/stream-decorator-reference.html#outerHashJoin">Stream Decorator Reference: outerHashJoin</a>
+     * @param leftStream
+     * @param rightStream
+     * @param onClause
+     * @return
+     */
+    public static StreamExpression outerHashJoin(StreamExpression leftStream, Hashed rightStream, On onClause) {
+        return new StreamExpression("outerHashJoin")
+            .withParameter(leftStream)
+            .withParameter(rightStream)
+            .withParameter(onClause);
+    }
+
+}
